requirements.txt
SpeechRecognition--3.9.0